requirements.txt
telethon>=1.36.0
python-telegram-bot>=21.6
python-dotenv>=1.0.1
requests>=2.32.0
numpy>=1.26.0
